refactor: ComputerNumberSet 과 UserNumberSet 으로 분리
