+++ src/main/java/baseball/Application.java
@@ -1,14 +1,15 @@
 package baseball;
 
 import controller.GameController;
-import model.NumberSet;
+import model.ComputerNumberSet;
+import model.UserNumberSet;
 import view.TerminalView;
 
 public class Application {
 	public static void main(String[] args) {
 		TerminalView terminalView = new TerminalView();
-		NumberSet computerNumberSet = new NumberSet();
-		NumberSet userNumberSet = new NumberSet();
+		ComputerNumberSet computerNumberSet = new ComputerNumberSet();
+		UserNumberSet userNumberSet = new UserNumberSet();
 		GameController gameController = new GameController(terminalView, computerNumberSet, userNumberSet);
 		gameController.startGame();
 	}
